backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
graphene-django>=3.0
django-graphql-jwt>=0.4
Pillow>=10.0.0
celery>=5.3
channels[daphne]>=4.0
channels_redis>=4.1
django-celery-results>=2.5
django-channels-graphql-ws>=1.0.0rc7
pytest>=7.0
pytest-django>=4.0
